refactor: Restructure project files for better organization, readability, and maintainability
