+++ cmd/config/config.go
@@ -17,7 +17,7 @@ limitations under the License.
 package config
 
 import (
-	"github.com/jkaninda/goma-gateway/internal/logger"
+	"github.com/jkaninda/goma-gateway/pkg/logger"
 	"github.com/spf13/cobra"
 )
