+++ cmd/config/init.go
@@ -16,8 +16,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 import (
-	"github.com/jkaninda/goma-gateway/internal/logger"
-	"github.com/jkaninda/goma-gateway/pkg"
+	"github.com/jkaninda/goma-gateway/internal"
+	"github.com/jkaninda/goma-gateway/pkg/logger"
 	"github.com/spf13/cobra"
 )
