+++ internal/middleware/middleware.go
@@ -18,7 +18,7 @@ limitations under the License.
 import (
 	"encoding/base64"
 	"encoding/json"
-	"github.com/jkaninda/goma-gateway/internal/logger"
+	"github.com/jkaninda/goma-gateway/pkg/logger"
 	"io"
 	"net/http"
 	"net/url"
