+++ internal/var.go
